Refactoring TransactionDecoder
