+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/core/encoding/DecodingContext.java
@@ -1,0 +1,6 @@
+package org.hyperledger.besu.ethereum.core.encoding;
+
+enum DecodingContext {
+  INTERNAL,
+  NETWORK,
+}
