+++ ethereum/core/src/test/java/org/hyperledger/besu/ethereum/core/encoding/TransactionRLPEncoderTest.java
@@ -37,7 +37,7 @@ class TransactionRLPEncoderTest {
   @Test
   void encodeFrontierTxNominalCase() {
     final Transaction transaction =
-        TransactionDecoder.decodeForWire(RLP.input(Bytes.fromHexString(FRONTIER_TX_RLP)));
+        TransactionDecoder.decode(RLP.input(Bytes.fromHexString(FRONTIER_TX_RLP)));
     final BytesValueRLPOutput output = new BytesValueRLPOutput();
     TransactionEncoder.encodeForWire(transaction, output);
     assertThat(output.encoded().toHexString()).isEqualTo(FRONTIER_TX_RLP);
@@ -46,7 +46,7 @@ class TransactionRLPEncoderTest {
   @Test
   void encodeEIP1559TxNominalCase() {
     final Transaction transaction =
-        TransactionDecoder.decodeForWire(RLP.input(Bytes.fromHexString(EIP1559_TX_RLP)));
+        TransactionDecoder.decode(RLP.input(Bytes.fromHexString(EIP1559_TX_RLP)));
     final BytesValueRLPOutput output = new BytesValueRLPOutput();
     TransactionEncoder.encodeForWire(transaction, output);
     assertThat(output.encoded().toHexString()).isEqualTo(EIP1559_TX_RLP);
@@ -70,8 +70,7 @@ class TransactionRLPEncoderTest {
   @Test
   void shouldEncodeWithHighNonce() {
     final Transaction transaction =
-        TransactionDecoder.decodeForWire(
-            RLP.input(Bytes.fromHexString(NONCE_64_BIT_MAX_MINUS_2_TX_RLP)));
+        TransactionDecoder.decode(RLP.input(Bytes.fromHexString(NONCE_64_BIT_MAX_MINUS_2_TX_RLP)));
     final BytesValueRLPOutput output = new BytesValueRLPOutput();
     TransactionEncoder.encodeForWire(transaction, output);
     assertThat(output.encoded().toHexString()).isEqualTo(NONCE_64_BIT_MAX_MINUS_2_TX_RLP);
